Removed Copy-On-Write Functionality
Removed COW functionality from the stream layer.
