+++ runtime/src/main/java/org/corfudb/runtime/view/stream/AbstractQueuedStreamView.java
@@ -1,6 +1,7 @@
 package org.corfudb.runtime.view.stream;
 
 import lombok.Data;
+import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.corfudb.protocols.logprotocol.CheckpointEntry;
 import org.corfudb.protocols.wireprotocol.DataType;
@@ -45,8 +46,23 @@ import java.util.stream.Collectors;
  * <p>Created by mwei on 1/6/17.
  */
 @Slf4j
-public abstract class AbstractQueuedStreamView extends
-        AbstractContextStreamView {
+public abstract class AbstractQueuedStreamView implements IStreamView, AutoCloseable {
+
+    /**
+     * The ID of the stream.
+     */
+    @Getter
+    private final UUID id;
+
+    /**
+     * The runtime the stream view was created with.
+     */
+    @Getter
+    private final CorfuRuntime runtime;
+
+    // Context of this stream
+    @Getter
+    private final StreamContext context;
 
     final StreamOptions options;
 
@@ -55,12 +71,150 @@ public abstract class AbstractQueuedStreamView extends
      * @param streamId  The ID of the stream
      * @param runtime   The runtime used to create this view.
      */
-    public AbstractQueuedStreamView(final CorfuRuntime runtime,
-                                    final UUID streamId,
+    public AbstractQueuedStreamView(final CorfuRuntime runtime, final UUID streamId,
                                     StreamOptions options) {
-        super(runtime, streamId);
+        this.runtime = runtime;
+        this.id = streamId;
+        this.context = new StreamContext(id, Address.MAX);
         this.options = options;
     }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void reset() {
+        context.reset();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void seek(long globalAddress) {
+        // now request a seek on the context
+        this.context.seek(globalAddress);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public final ILogData nextUpTo(final long maxGlobal) {
+        // Don't do anything if we've already exceeded the global pointer.
+        if (context.getGlobalPointer() > maxGlobal) {
+            return null;
+        }
+
+        // Get the next entry from the underlying implementation.
+        final ILogData entry = getNextEntry(getContext(), maxGlobal);
+
+        if (entry != null) {
+            // Update the pointer.
+            updatePointer(entry);
+        }
+
+        // Return the entry.
+        return entry;
+    }
+
+    /** {@inheritDoc}
+     */
+    @Override
+    public final List<ILogData> remainingUpTo(long maxGlobal) {
+        final List<ILogData> entries = getNextEntries(getContext(), maxGlobal);
+
+        // Nothing read, nothing to process.
+        if (entries.size() == 0) {
+            // We've resolved up to maxGlobal, so remember it. (if it wasn't max)
+            if (maxGlobal != Address.MAX) {
+                // Set Global Pointer and check that it is not pointing to an address in the trimmed space.
+                getContext().setGlobalPointerCheckGCTrimMark(maxGlobal);
+            }
+            return entries;
+        }
+
+        // Otherwise update the pointer
+        if (maxGlobal != Address.MAX) {
+            // Set Global Pointer and check that it is not pointing to an address in the trimmed space.
+            getContext().setGlobalPointerCheckGCTrimMark(maxGlobal);
+        } else {
+            // Update pointer from log data and then validate final position of the pointer against GC trim mark.
+            updatePointer(entries.get(entries.size() - 1));
+            getContext().validateGlobalPointerPosition(getCurrentGlobalPosition());
+        }
+
+        // And return the entries.
+        return entries;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean hasNext() {
+        return getHasNext(getContext());
+    }
+
+
+    /** Retrieve the next entry in the stream, given the context.
+     *
+     * @param context       The context to retrieve the next entry from.
+     * @param maxGlobal     The maximum global address to read to.
+     * @return              Next ILogData for this context
+     */
+
+
+    /** Retrieve the next entries in the stream, given the context.
+     *
+     * <p>This function is designed to implement a bulk read. In a bulk read,
+     * one of the entries may cause the context to change - the implementation
+     * should check if the entry changes the context and stop reading
+     * if this occurs, returning the entry that caused contextCheckFn to return
+     * true.
+     *
+     * <p>The default implementation simply calls getNextEntry.
+     *
+     * @param context           The context to retrieve the next entry from.
+     * @param maxGlobal         The maximum global address to read to.
+     * @return                  A list of the next entries for this context
+     */
+    protected List<ILogData> getNextEntries(StreamContext context, long maxGlobal) {
+        final List<ILogData> dataList = new ArrayList<>();
+        ILogData thisData;
+
+        while ((thisData = getNextEntry(context, maxGlobal)) != null) {
+            // Add this read to the list of reads to return.
+            dataList.add(thisData);
+
+            // Update the pointer, because the underlying implementation
+            // will expect it to be updated when we call getNextEntry() again.
+            updatePointer(thisData);
+        }
+
+        return dataList;
+    }
+
+    /** Update the global pointer, given an entry.
+     *
+     * @param data  The entry to use to update the pointer.
+     */
+    private void updatePointer(final ILogData data) {
+        // Update the global pointer, if it is non-checkpoint data.
+        if (data.getType() == DataType.DATA && !data.hasCheckpointMetadata()) {
+            // Note: here we only set the global pointer and do not validate its position with respect to the trim mark,
+            // as the pointer is expected to be moving step by step (for instance when syncing a stream up to maxGlobal)
+            // The validation is deferred to these methods which call it in advance based on the expected final position
+            // of the pointer.
+            getContext().setGlobalPointer(data.getGlobalAddress());
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void close() {}
 
     /** Add the given address to the resolved queue of the
      * given context.
@@ -77,11 +231,13 @@ public abstract class AbstractQueuedStreamView extends
         }
     }
 
-    /**
-     * {@inheritDoc}
+     /** Retrieve the next entry in the stream, given the context.
+     *
+     * @param context       The context to retrieve the next entry from.
+     * @param maxGlobal     The maximum global address to read to.
+     * @return              Next ILogData for this context
      */
-    @Override
-    protected ILogData getNextEntry(StreamContext context,
+    private ILogData getNextEntry(StreamContext context,
                                     long maxGlobal) {
         // If we have no entries to read, fill the read queue.
         // Return if the queue is still empty.
@@ -490,25 +646,21 @@ public abstract class AbstractQueuedStreamView extends
         }
     }
 
-    /**
-     * {@inheritDoc}
+    /** Return whether calling getNextEntry() may return more
+     * entries, given the context.
+     * @param context       The context to retrieve the next entry from.
+     * @return              True, if getNextEntry() may return an entry.
+     *                      False otherwise.
      *
      * <p> We indicate we may have entries available
      * if the read queue contains entries to read -or-
      * if the next token is greater than our log pointer.
      */
-    @Override
-    public boolean getHasNext(StreamContext context) {
+    private boolean getHasNext(StreamContext context) {
         return  !context.readQueue.isEmpty()
                 || runtime.getSequencerView().query(context.id).getToken().getSequence()
                 > context.getGlobalPointer();
     }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void close() {}
 
     /**
      * {@inheritDoc}
@@ -753,6 +905,11 @@ public abstract class AbstractQueuedStreamView extends
         return getContext().getGlobalPointer();
     }
 
+    @Override
+    public String toString() {
+        return Utils.toReadableId(context.id) + "@" + getContext().getGlobalPointer();
+    }
+
     /**
      * Represents a contained form of a stream's checkpoint.
      * Only with relevant information for the stream view.
@@ -826,5 +983,4 @@ public abstract class AbstractQueuedStreamView extends
             numEntries += entries;
         }
     }
-
 }
